feat: formatação e ajuste de tipos
