+++ src/services/course-service.ts
@@ -1,30 +1,6 @@
 import { apiClient } from '@/lib/api-client'
+import { TCourse } from '@/types/course'
 import { API_CONFIG } from '@/utils/api-config'
-
-export type TCourse = {
-  id: number
-  title: string
-  slug: string
-  status: string
-  short_description: string
-  thumbnail: string
-  banner: string
-  banner_mobile: string
-  long_description: string
-  level: string
-  featured: boolean
-  store: {
-    id: number
-    name: string
-    slug: string
-  }
-  teachers: Array<{
-    id?: number
-    name?: string
-    avatar?: string
-  }>
-  visibility: string
-}
 
 const isClient = typeof window !== 'undefined'
